home-manager-tool: Depricate home-manager paths
Depricate hard-coded fallback home-manager paths. Specificly
"${XDG_CONFIG_HOME:-$HOME/.config}/nixpkgs/home-manager" and
"$HOME/.nixpkgs/home-manager".

Use @HOME_MANAGER_PATH@ if it has been provided and points to something
that exisits. Warn the user if it does not point to something.

If we have not been provided with a @HOME_MANAGER_PATH@ that exisits,
then for both hard-coded paths show a warning if something exisits where
the paths are pointing.

This no longer attempts to use either of the hard-coded paths as
fallbacks for the home-manager path.